Import of GTERMARS/Alien-scriptaculous-1.7.1.2_03 from CPAN.
gitpan-cpan-distribution: Alien-scriptaculous
gitpan-cpan-version:      1.7.1.2_03
gitpan-cpan-path:         GTERMARS/Alien-scriptaculous-1.7.1.2_03.tar.gz
gitpan-cpan-author:       GTERMARS
gitpan-cpan-maturity:     developer

+++ t/02-install.t
@@ -1,21 +1,38 @@
 use strict;
 use warnings;
-use Test::More tests => 3;
+use Test::More;
 use Test::Exception;
 use File::Path qw(rmtree);
 use Alien::scriptaculous;
 
 my $dir = 't/eraseme';
 
-# Do an install and make sure that at least one file from each of the 'lib' and
-# 'src' directories was installed properly.
-Alien::scriptaculous->install( $dir );
-foreach my $file (qw( prototype.js scriptaculous.js )) {
-    ok( -e "$dir/$file", "$dir/$file exists" );
+# Figure out how many tests we're going to run
+my @files = Alien::scriptaculous->files();
+plan tests => 2 + scalar @files;
+
+# Clean up from any previous test run.
+cleanup_old_test_run: {
+    rmtree( $dir ) if (-e $dir);
 }
 
-# Re-install into the same directory, to make sure that it doesn't choke.
-lives_ok { Alien::scriptaculous->install($dir) };
+# Install, and make sure that all of the files got installed properly.
+install_scriptaculous: {
+    Alien::scriptaculous->install( $dir );
+    # make sure Prototype got installed
+    ok( -e "$dir/prototype.js", "prototype installed" );
+    # make sure our files got installed
+    foreach my $file (@files) {
+        ok( -e "$dir/$file", "$file exists" );
+    }
+}
 
-# Clean out the test directory
-rmtree( $dir );
+# Install on top of an existing install; shouldn't choke
+reinstall_scriptaculous: {
+    lives_ok { Alien::scriptaculous->install($dir) };
+}
+
+# Clean up after ourselves
+cleanup: {
+    rmtree( $dir );
+}
